+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make sure Mapbox and Maplibre attributions don't overlap the styles dropdown
 - Add support for MapTiler SDK
 - Consistify URL normalization between Mapbox and Maptiler handling
+- Fix bug that did not allow valid bearing values to be entered
 
 ## 0.16.0