+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Consistify URL normalization between Mapbox and Maptiler handling
 - Fix validation for style preset urls not allowing multi level dropdown syntax
 - Fix bug that did not allow valid bearing values to be entered
+- Fix validation for style preset urls not allowing multi level dropdown syntax
 - Update Maplibre GL to `^5.2.0`
 - Update Mapbox GL to `^3.10.0`
 - Update Maptiler SDK to `^3.0.1`