+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Update Maplibre GL to `^5.2.0`
 - Update Mapbox GL to `^3.10.0`
 - Update Maptiler SDK to `^3.0.1`
+- Fix fetch url transform of specific protocols
+- Fix screenshotting with Maplibre update
 
 ## 0.16.0